refactor(notification): 알림 렌더 전략이 payload 생성 책임을 갖도록 변경
actor-post 기반 알림의 추상 전략을 제거하고 각 concrete strategy가 리소스 조회, payload 생성, 썸네일 해석 흐름을 직접 드러내도록 정리합니다.
