+++ src/main/kotlin/com/techtaurant/mainserver/notification/application/PostCommentNotificationListItemRenderStrategy.kt
@@ -1,26 +1,62 @@
 package com.techtaurant.mainserver.notification.application
 
+import com.techtaurant.mainserver.notification.dto.NotificationListItemResponse
+import com.techtaurant.mainserver.notification.enums.NotificationTargetType
 import com.techtaurant.mainserver.notification.enums.NotificationType
-import com.techtaurant.mainserver.post.entity.Post
 import com.techtaurant.mainserver.user.entity.User
 import org.springframework.stereotype.Component
 import java.util.UUID
 
 @Component
 internal class PostCommentNotificationListItemRenderStrategy(
-    notificationPayloadService: NotificationPayloadService,
-    notificationPayloadResourceResolver: NotificationPayloadResourceResolver,
-) : AbstractActorPostNotificationListItemRenderStrategy(
-        notificationPayloadService = notificationPayloadService,
-        notificationPayloadResourceResolver = notificationPayloadResourceResolver,
-    ) {
+    private val notificationPayloadService: NotificationPayloadService,
+    private val notificationPayloadResourceResolver: NotificationPayloadResourceResolver,
+) : NotificationListItemRenderStrategy {
     override val type: NotificationType = NotificationType.POST_COMMENT
-    override val messageKey: String = "notification.payload.post-comment"
 
-    override fun selectThumbnailMedia(
+    override fun render(commands: List<NotificationListItemRenderCommand>): Map<UUID, NotificationListItemResponse> {
+        if (commands.isEmpty()) {
+            return emptyMap()
+        }
+
+        val actorUserIds = commands.mapNotNull { it.arguments.findTargetId(NotificationTargetType.USER) }
+        val actorsById = notificationPayloadResourceResolver.findUsersById(actorUserIds)
+        val actorProfileImageUrlByUserId =
+            notificationPayloadResourceResolver.resolveActorProfileImageUrlByUserId(actorsById)
+
+        val postIds = commands.mapNotNull { it.arguments.findTargetId(NotificationTargetType.POST) }
+        val postsById = notificationPayloadResourceResolver.findPostsById(postIds)
+
+        return commands.associate { command ->
+            val recipient = command.recipient
+            val notificationId = recipient.notification.id!!
+            val actor = command.arguments.findTargetId(NotificationTargetType.USER)?.let(actorsById::get)
+            val post = command.arguments.findTargetId(NotificationTargetType.POST)?.let(postsById::get)
+            val media = createMedia(actor, actorProfileImageUrlByUserId)
+
+            notificationId to
+                NotificationListItemResponse.from(
+                    recipient = recipient,
+                    payloadHtml =
+                        notificationPayloadService.buildPayload(
+                            messageKey = MESSAGE_KEY,
+                            messageArguments = listOf(actor?.name.orEmpty(), post?.title.orEmpty()),
+                        ),
+                    thumbnailUrl = notificationPayloadService.resolveThumbnailUrl(media),
+                    arguments = command.arguments,
+                )
+        }
+    }
+
+    private fun createMedia(
         actor: User?,
-        post: Post?,
         actorProfileImageUrlByUserId: Map<UUID, String>,
-        postThumbnailUrlByPostId: Map<UUID, String>,
-    ): NotificationPayloadService.NotificationPayloadMedia = createActorProfileMedia(actor, actorProfileImageUrlByUserId)
+    ): NotificationPayloadService.NotificationPayloadMedia =
+        NotificationPayloadService.NotificationPayloadMedia(
+            url = notificationPayloadResourceResolver.resolveActorProfileImageUrl(actor, actorProfileImageUrlByUserId),
+        )
+
+    private companion object {
+        const val MESSAGE_KEY = "notification.payload.post-comment"
+    }
 }
